refactor: 고객지원 관련 컴포넌트 리팩토링
