+++ src/app/support/page.tsx
@@ -1,13 +1,27 @@
-import SupportSkeleton from "@/components/skeleton/support/SupportSkeleton";
-import SupportHeaderContainer from "@/containers/support/SupportHeaderContainer";
-import { Suspense } from "react";
+import SupportHeader from "@/components/support/SupportHeader";
+import SupportAbout from "@/components/support/about/SupportAbout";
+import SupportFAQList from "@/components/support/faq/SupportFAQList";
+import SupportTerms from "@/components/support/terms/SupportTerms";
+import SupportNoticeList from "@/components/support/notice/SupportNoticeList";
 
-export default function Page() {
+export default async function Page({
+  searchParams,
+}: {
+  searchParams: Promise<{ [key: string]: string | undefined }>;
+}) {
+  const menu = (await searchParams).menu;
+
+  if (menu === undefined) {
+    throw Error("Not Found");
+  }
+
   return (
     <div className="mb-8 w-full">
-      <Suspense fallback={<SupportSkeleton />}>
-        <SupportHeaderContainer />
-      </Suspense>
+      <SupportHeader menu={menu} />
+      {menu === "about" && <SupportAbout />}
+      {menu === "notice" && <SupportNoticeList />}
+      {menu === "faq" && <SupportFAQList />}
+      {menu === "terms" && <SupportTerms />}
     </div>
   );
 }
