+++ src/components/support/SupportHeader.tsx
@@ -1,27 +1,21 @@
+import { SUPPORT_TABS } from "@/constants/support/supportTabs";
 import Link from "next/link";
 
 interface SupportHeaderProps {
-  active: string | null;
+  menu: string;
 }
 
-const tabs = [
-  { name: "서비스소개", value: "about" },
-  { name: "공지사항", value: "notice" },
-  { name: "FAQ", value: "faq" },
-  { name: "이용약관", value: "terms" },
-];
-
-const SupportHeader = ({ active }: SupportHeaderProps) => {
+const SupportHeader = ({ menu: active }: SupportHeaderProps) => {
   return (
     <div className="mb-4 flex w-full flex-col items-start py-1">
-      <div className={"flex w-full flex-col items-start gap-y-4 pt-[2.375rem]"}>
-        <h1 className={"w-full text-start text-[1.75rem] font-bold text-black"}>
+      <div className="flex w-full flex-col items-start gap-y-4 pt-[2.375rem]">
+        <h1 className="w-full text-start text-[1.75rem] font-bold text-black">
           고객지원
         </h1>
-        <p className={"text-gray1"}> 솔리투어 고객지원에서 도와드릴게요 </p>
+        <p className="text-gray1">솔리투어 고객지원에서 도와드릴게요</p>
       </div>
       <article className="flex max-w-full flex-wrap justify-start gap-x-1 gap-y-2 pt-12">
-        {tabs.map((tab) => (
+        {SUPPORT_TABS.map((tab) => (
           <Link
             key={tab.value}
             href={`/support?menu=${tab.value}`}
