+++ src/components/support/terms/SupportTerms.tsx
