+++ src/constants/support/supportTabs.ts
@@ -1,0 +1,6 @@
+export const SUPPORT_TABS = [
+  { name: "서비스 소개", value: "about" },
+  { name: "공지사항", value: "notice" },
+  { name: "FAQ", value: "faq" },
+  { name: "이용약관", value: "terms" },
+] as const;
